critic-verse/backend/requirements.txt
# backend
fastapi == 0.103.2
uvicorn == 0.23.2
python-dotenv == 1.0.0
elasticsearch==8.10.0
elastic-transport==8.4.0